refactor: jpa 인프라 레벨로 옮기기
