+++ src/main/java/com/example/stormpaws/infra/jpa/UserRepository.java
@@ -1,12 +1,13 @@
-package com.example.stormpaws.repository;
+package com.example.stormpaws.infra.jpa;
 
+import com.example.stormpaws.domain.IRepository.IUserRepository;
 import com.example.stormpaws.domain.model.UserModel;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserModel, String> {
+public interface UserRepository extends JpaRepository<UserModel, String>, IUserRepository {
   Optional<UserModel> findByOauthTypeAndOauthId(String oauthType, String oauthId);
 
   Optional<UserModel> findByEmail(String email);
