+++ src/main/java/com/example/stormpaws/service/UserService.java
@@ -1,7 +1,7 @@
 package com.example.stormpaws.service;
 
+import com.example.stormpaws.domain.IRepository.IUserRepository;
 import com.example.stormpaws.domain.model.UserModel;
-import com.example.stormpaws.repository.UserRepository;
 import com.example.stormpaws.service.OAuth.IOAuthProvider;
 import com.example.stormpaws.service.OAuth.OAuthProviderFactory;
 import com.example.stormpaws.service.dto.AuthDataDTO;
@@ -15,12 +15,12 @@ import org.springframework.stereotype.Service;
 public class UserService {
 
   private final OAuthProviderFactory oauthProviderFactory;
-  private final UserRepository userRepository;
+  private final IUserRepository userRepository;
   private final ITokenProvider tokenProvider;
 
   public UserService(
       OAuthProviderFactory oauthProviderFactory,
-      UserRepository userRepository,
+      IUserRepository userRepository,
       ITokenProvider tokenProvider) {
     this.oauthProviderFactory = oauthProviderFactory;
     this.userRepository = userRepository;
