+++ util/rest_client/user/GET token refresh.http
@@ -2,4 +2,4 @@ GET {{stormpaw-host}}/user/token-refresh/
 Content-Type: application/json
 Authorization: Bearer {{stormpaw-jwt}}
 
-# acees 토큰 만료시 새로운 access 토큰 및 refresh 토큰을 보내줌.
+# access 토큰 만료시 새로운 access 토큰을 발급해줍니다.
